딥러닝과제/[HW1]/requirements.txt
requests>=2.31.0,<3
PyYAML>=6.0.2,<7


